requirements.txt
py-clob-client==0.34.6
python-telegram-bot==22.7
aiohttp==3.11.14
aiohttp-socks==0.10.1
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.25.3
